[+FEATURE] TYPO3: Further adapt TYPO3 to the new content repository
A snapshot of my work in progress. Mainly contains the refactoring
of the TypoScript part which features a generic Node TypoScript
object instead of specific Text, Headline etc. objects.

Another big part is the NodeController and its related services.

In order to use the NodeController in an ExtJS context, an
Object Converter for nodes needed to be implemented.

Change-Id: Ied9af0ac9926ec1d509fc3cf0d0c65e8d9c2e781