[FEATURE] TYPO3 (UI): Make Aloha work with non-selectable areas
The behavior is now a lot more like described by the design:
If you click twice on the page, the Edit Mode is activated,
which deactivates everything not being editable.

The second double-click really activates aloha then. It is not possible
right now to stay in edit mode, but disable aloha.

Additionally, some cleanups were happening in this change, I
especially renamed <typo3:editable> to <typo3:aloha.editable>.

Change-Id: Ic2cf996dc255df69496109ec565e0216811b6eac