[+FEATURE] TYPO3: Automatically create text element on new page
This enhances the ExtDirect NodeController so that whenever a new node
of type TYPO3:Page is created, also a new section containing a new
text mode with some dummy text is added.

Currently then name of the section is hardcoded ("main"). A future
implementation should check which Fluid template is currently active
for the Page, determine the section names and create a text element
in each of them.

Change-Id: I290a15c9cfbbe7ec97dea9bc62bb3f51b37429b5
Resolves: #11200