[BUGFIX] TYPO3 (UI): Make Aloha activation with double clicks
This change introduces the correct double-click
behavior for Aloha. Right now, this is not directly
built into Aloha core; instead, we re-wire certain
events and handle them on our own.

The code is not really easy to understand, as there are
lots of events which need to be fired in the correct order.
Because of this, longer inline comments are used.

Change-Id: I28e951863c87cbddb60170fa0981cb3d79643576