[TASK] Cleanups and improvements in TYPO3 UI
* The current workspace is now stored in a Session.

This makes it impossible to tamper with the currently active workspace
if not allowed to.
Furthermore, no URLs contain the workspace anymore, making the RDF
annotations we use more semantically correct.

* No __context in JavaScript needed anymore

We do not need the __context in JavaScript anymore, instead we can just
pass the nodePath as string around.
Due to incapabilities of the current property mapper, we artificially
have to wrap the node path into {__nodePath: nodePath} if we want to
send it to the server. However, there are TODOs added, so once the new
property mapper is in place, we can further cleanup this.

* VIE Bugfix

This change contains a VIE bugfix, https://github.com/bergie/VIE/pull/13

* Further decoupling on the JavaScript side

The Content Editor "Save" button implemented in SaveContentContentDialog
now listens to events, reducing the dependencies between Frontend Core
and Backend.

* Improved CGL and doc comments along the way.

* Many JavaScript bugfixes

This includes the BreadcrumbMenu, Drag&Drop functionality, Page Position
Selector, Management View, among others.

* Manual functional tests

The backend has been tested under Firefox 4 for the following cases

- Edit Content (OK)
- Edit HTML (OK)
- Add Page, with Position Selection (OK)
- Move Page (OK)
- Delete Page (OK)
- Management View - Tree (OK)
- Management View - Content Pane (OK)
- Add Content (works 90%)
-> it is not possible yet to add a new element after an HTML element,
   this needs to be solved concept-wise.
-> placeholders when content is empty do not fully work.
- Move Content (works 90%)

-> Because moving a node touches all of the nodes between the source and
   the target node, this sometimes has the effect of placing a node at
   the wrong position.

Furthermore, all of the above cases have been tested for Workspace
compatibility (i.e. that they only show up in live workspace after
publishing).

* MISSING: Unit Tests / JavaScript tests have to be adjusted.

Change-Id: Ib9ac9c97cc03c148e5783913b0c51d9be3dd40d2