[+BUGFIX] Re-throw StopActionException in Plugin TS Object
Since change I224cf3dbc1a0290ab29cb7bbbb8651f275ef9f32
all exceptions that are thrown in a sub request dispatching
process are caught and logged in the Plugin TypoScript object.
This included StopActionExceptions. But they have to be
re-thrown in order to be able to handle them in the top level request

Change-Id: I914985baacd7d44d9e223a9dee9a77191fa07f58