[TASK] Introduce a NodeInterface
The relation between Node and ProxyNode was too weak: if one forgets to
implement a feature in ProxyNode which has been introduced to Node, the
workspaces functionality might show some unexpected behavior which can't
be spotted easily. With this change ProxyNode does not inherit from Node
anymore but both classes share the same interface.

Change-Id: Ic442e42000ef885b8c3351aa5cf7b535bf693673