[TASK] New approach for frontend routing and context handling
This splits up the routing and controller for the frontend and REST
service so we have better control on each of them. The REST service
is now protected by a policy (requires admin rights) and the required
workspace needs to be passed as in a special "context node path".
If no workspace is specified, "live" is assumed.

A context node path looks like this (workspace = "user-admin"):
  /sites/typo3org/homepage/about@user-admin.html
The UI (mainly WebsiteContainer) now uses this kind of URI for the
website iFrame. That iFrame now uses the new Frontend\NodeController
instead of the REST service.

This change set also introduces a new "User" object which has
"UserPreferences". These preferences are meant for storing the
currently selected workspace ("context.workspace") but can also be
used for other purposes later on.

Also contains cosmetic changes to the RoutingLoggingAspeect class and
route examples.

Change-Id: Ib7fb8e66f1a8ef75cdea3e82c7b56f3fb2113f09