[TASK] Introduce consistent context concept
This streamlines existing code which somehow deals with (Content)
Context and introduces a consistent vocabulary and format for context
references.

The NodeObjectConverter now returns a string (an absolute context node
path) instead of the actual Node object.

Node Path
A path pointing to a node. A Node Path can be absolute (leading forward
slash) or relative  (usually to the current Site Node). If nothing else
is specified, the node path is assumed to be absolute.

Context Node Path
A node path which contains additional context information, such as the
workspace name. The Absolute Context Node Path marks the absolute
position and thus starts with a forward slash. A Relative Context Node
Path does not start with a slash and is relative to another node,
usually the Site Node. If nothing else is specified, a context node path
is assumed to be absolute.

Node Identifier
UUID of the node. This identifier is only unique within a single
workspace and is also used to find the corresponding node in another
workspace.

Site Node
The root node of a site, located below /Sites.

Change-Id: I81ab3d5f94f59fd123ad70b8bef9c31cb80bea5b