[TASK] Add rudimentary varnish support
This lets the frontend's NodeController send two cache headers which
at least allow for some basic caching by varnish. For now TYPO3 will
still initiate a session regardless of you're logged in or not and thus
set a cookie. This must be handled (for now) by some quick fix regex
in the Varnish vcls ... 

Change-Id: I8e20e5bfe20fa7aed14058a406aaa79d0d17832d