[TASK] Fix ExtDirect calls in with vendor-namespaces
It wasn't possible to use TYPO3.TYPO3.Controller.LoginController format,
so I had to use the underscores. To prevent using dots and underscore
in one name (like it was before) I removed the namespace of the provider
by default.

Otherwise we would end up with calls like:
TYPO3.TYPO3.TYPO3_TYPO3_Controller_LoginController

Might still need to be changed later on, but it works for now.

Change-Id: Ie016bc40587563152551675dd4bfe0dafe84cb62