[BUGFIX] Avoid error in object-to-JSON serialization
When the ContentElementWrappingService tried to convert an object
to JSON but the value was not an object, conversion would fail.
The change skips serialization if the value at hand is not an object.

Can be reproduced by loading a page with "text with image" elements
that have no image yet.

Change-Id: Ia6f0a726eca063098426aea12b2fec0f98b96bbf