[TASK] Text changes trigger autosaving of page
Text changes trigger autosaving as long as you type. If the
connection is down it will be shown at the publish button
which also is disabled in that case.

There are no concurrent save requests if a request is running
while changes happen the next save request will be postponed.

Change-Id: I5b267af190f639dc8fccdef4ea0aebab27905bea