[BUGFIX] Fix backend login
The change in https://review.typo3.org/#change,2561 introduced a
regression which made logging into the backend impossible.

Security\Context::separateActiveAndInactiveTokens marked the Typo3BackendProvider
as *inactive*, because the request pattern did not match the login controller.

It is strange that this error did not occur while I tested the change,
but presumably I forgot to clear the browser's cookies in this case, not
starting with a fresh browser session.

When testing this change, make sure to clear your browser's cookies before.

Change-Id: Icae2e326f71b334a516520e2b9e82516b50647e1