[FEATURE] Shortcut TS object
This change adds a Shortcut TS object (TYPO3.TYPO3:Shortcut) that
for now redirects to it's first child page. The menu and link/uri
node view helpers also resolve shortcuts through the frontend node
routepart handler.

Shortcuts are resolved recursively, that is if the first subpage
is again a shortcut, it's first child node is tried.

Some whitespace fixes to the Page class along the way.

Change-Id: I9a12b8a05044e57d0d3106e426d5f1a57859cce3
Resolves: #33723