[BUGFIX] Fixed duplicated content which is shown on first call
I finally understood the problem of why Phoenix sometimes rendered duplicate
content: It is because the ContentElement ViewHelper (which wraps the
content element with the surrounding div) was of type *singleton*:

In this case, two invocations of this ViewHelper would have the same child nodes
set, leading to the bogus rendering.

This ViewHelper violated the basic rule for singleton ViewHelpers, which must
be completely *stateless*, i.e. depend only on the parameters of the render()
method. However, it also (implicitly) accessed the sub nodes by calling
$this->renderChildren() -- making it effectively a stateful ViewHelper.

Change-Id: I1cce6d80f040d0cef95146b0aaacd4ca0c9ecda6