[TASK] modify headline in text with headline from h2 to h1 as it was before
This is just a quick fix to correct the rendering on the flow3.org website,
so we need to change the h2 back to the h1 as it was before the refactored
rendering.

Lateron, we need to discuss whether a content element even has a headline,
or make the headline level configurable.

Change-Id: I395b8410ec3f263917e0290ed3963d1bdb853da8