[BUGFIX] (UI): Fix workspace updates when live content is modified
If a content element which is inside the live workspace is being modified
through the user interface, the content element is effectively copied
to the user's workspace when the change is saved on the server-side.

When this happens, we need to update the workspace name in the JavaScript
model such that the UI can react appropriately, i.e. mark the node as
"Publishable".

To reproduce, take a plain Phoenix instance (with no nodes inside the
user workspace) and modify some content. Then, watch how the contents
state in the bottom breadcrumb menu changes to "modified" and then
back to "unmodified". After applying this patch, it should change to
"modified" and then to "publishable".

Related: #PHX-54
Change-Id: Iecfd55a4198f2de25b79db99dc6de73ccd80f50e