[BUGFIX] NodeRoutePartHandler should use getNode without context
The NodeRoutePartHandler tries to use getNode with a context
node path in case we are not in live workspace. That won't work
as getNode expects the pure node path without context.

This change fixes the problem so that the current context is
used instead. This takes care for example the exception in the
registration plugin in phoenix demo site when logged in.

Change-Id: I5c4f8ecbc5cba4792d9dc160db0afdb59c67212b