[FEATURE] Make it visible an element can't be edited online on mouseover
This change adds an overlay over content elements which are not editable
inline on mouseover. Also introduces a primary action button in the
inspector panel. This primary action causes the overlay of the element
to have a button, double clicking this button fires the primary action from
within the inspector window.

Change-Id: I5caa12118ff67637aa460b6ac631c65b0ab340ea
Related: PHX-61