[BUGFIX] Fix issue with new icon over paste icon
When cutting/copying a content element, it's not possible
to insert the element because the paste icons are hidden
behind the new element icons.

At the moment the new icons are removed from block being
cut/copied, but all new icons should be hidden while
something is in the clipboard.

Related: PHX-22

Change-Id: I956adc1f4641543808d7907099ec13c8f839a55f