[BUGFIX] Fix JavaScript exception in ui.js caused by throw {}
There's an error in ui.js where a throw statement is given an
object instead of a string causing a fatal error.

Instead of throw statements ember_assert is now used for cleaner
and shorter code. (ember_assert() will change to Ember.Assert() in
the next version of Ember - v0.9.8.1+).

Change-Id: Id3275447132ce1001af402396314d41508b15535