[BUGFIX] Fix issue with popover buttons with dom elements inside
If a popover button has dom children, then the check if it's a button
being clicked doesn't work since it only check the click events target
dom for the class 'popover-button' and not it's parents. This issue
leads to popover-buttons having to be clicked twice since the popover
is hidden again after first click, since the hidePopover trigger is called.

Change-Id: Ifeef722a774e632a1bd91f6f884b9528044c4e80