[BUGFIX] Protected nodes can never be accessed
Due to a code fix in TYPO3CR a bug with fetching of access
protected nodes was uncovered that lead to non accessibility.
The problem was in Routing and PropertyMapping the
SecurityContext is not yet setup so roles are not available.

This change allows all nodes to be fetched from the TYPO3CR
during those early stages. So access protection must be checked
later (already implemented in the NodeController).

Related: #37267
Change-Id: If4e6bcaff73b136abf7435c19c49d31de73629dc