[!!!][BUGFIX] TypoScript date processor adjusted to correct time zone handling
With its former implementation the Date processor of TypoScript relied
strictly on the system's time zone, which is incorrect when
parsing Unix timestamps because these are per design GMT+00:00.

However, if desired, a timezone correction can be applied with a provided
setTimezone setter method.

This change is considered breaking because having relied on the behavior
(i.e. taking system timezone into account) would change the expectation now.

Change-Id: I1c34e3a8c213e59fb3a3c1ea6b33687411ba9d79
Fixes: #38544