[FEATURE] Button-Bar should be positioned reliably
We removed the old positioning of the button-bar; and after trying out jQuery UI
we found a CSS-only solution.

Additionally, the button-bar is correctly positioned inside sections; only
being visible when the section was selected or it is empty.

An active section is also outlined correctly with a border now.

This solution has been tested with Chrome and Firefox, for each of these sites:

* Old Demo site
* flow3.org site
* WIP demo site

Change-Id: Ia7301fc52f8c4e83c546f467b080f5fd8e19b977
Resolves: #40852