[FEATURE] Add content element overlay for not inline editables
This change re-introduces the content element overlay over elements
which are not inline editable. For now only elements which have no
inline editable property at all are overlayed.
The overlay is added before the contentelement-handle-bar, which
causes the handle-bar to be still clickable.

On single click the overlay tries to fire the click event on possible
elements with the class t3-primary-editor-action. If this element is
an Ember view and has a toggle() method it will also call that method.
This makes sure that for example the 'HTML Editor' toggle button will
be toggled.

Change-Id: I36434b54f02a2b4d40bd3aebf1860a18dfef313f
Resolves: #40761