[!!!][BUGFIX] Fix TypoScript template paths to start with uppercase char
The TypoScript files corresponding to the current node must reside in a
directory with the name of the node. Because node names are all
lowercase, the directory name also starts with a lowercase character.

This patch fixes this behavior and now requires the directories to start
with an uppercase character:

Before: Resources/Private/TypoScripts/home/about/About.ts2
After:  Resources/Private/TypoScripts/Home/About/About.ts2

Change-Id: I3b05cc4e04546c5a641f9d88592b3e13eaae4fe4
Resolves: #40933