[!!!][FEATURE] Improved TypoScripts directory structure and conventions
This changes the conventions for storing TypoScript files in a package's
resources directory.

By default, TypoScript files must be located in
"[package]/Resources/Private/TypoScripts". Below that base path, there
can be two directories: "Library" and "Nodes".

The first file to be included (if it exists) is "Library/Root.ts2".
This file may contain arbitrary TypoScript code and further inclusions.

Depending on the current node, further TypoScript files are loaded from
the "Nodes" directory. If the current node is "home/about-us/history",
the following additional files will be loaded, if they exist:

* TypoScripts/Nodes/Home.ts2
* TypoScripts/Nodes/Home/About-Us.ts2
* TypoScripts/Nodes/Home/About-Us/History.ts2

Although these directories and filenames are case insensitive, it is
recommended to use a natural mixed case instead of all lowercase.

Change-Id: Ib717bbb987edef32bb17ad790f7e2e6b76a3d94f
Resolves: #40936