[BUGFIX] Fix rendering tests
Due to the many TypoScript changes the rendering tests weren't
valid anymore. The test assumptions and TypoScript didn't match
the actual rendering process anymore. That was adapted.

Still the failing of the tests uncovers a flaw in current
rendering which is that headlines contain HTML instead of plain
text now and that this makes wrapping somewhat difficult.

Change-Id: I5850031b36edc35dfd35b59986d703c6620ac32c