[BUGFIX] while typing, the user should not loose the cursor position on autosave
Before this change, the user lost his cursor position after a successful autosave
(which happens automatically in the background).

The reason for this was as follows:
- after a save was complete, the VIE model was updated with server-side data
- this triggered a redraw of the VIE view, which is the content element
- this set the HTML contents of the editable using jQuery's html() function
- this, again, lead to the loss of the cursor position.

As bugfix, we mark model updates after autosaving as silent:

1) CreateJS had to be adjusted accordingly:
   https://github.com/bergie/create/commit/ffb40ffd723c700788ea08eb9b7f224eda9d134d
   https://github.com/bergie/create/commit/5177c5edb8c171d951a4150bc2d5e7aea722e59a

2) our own code which stores the workspace name of the content element has to do this
   in a silent way.

We not only updated create.js, but also VIE and hallo to be on a consistent version.

Furthermore, we switched to the un-minified version of these tools, as this makes
debugging more easy. These files will be minified by our build-system then.

Finally solved. :-)

Resolves: #41105
Change-Id: I408591e49ab7bad16bef641c81d4a781cc58c8da