[BUGFIX] Make clicking on new content element icons possible
Inside the new-content-element popover, clicking on the icon
of the content type did a redirect to "#", losing the current
page and showing the live workspace instead of adding a new
content element.

This change fixes this and makes the link easier to click and
fixing some styling issues as well as removing the non-used
HTML attributes as well.

Also removes the depreacted new content element template.

Change-Id: I39886ce9745e36f12c995b291f3f9a1acb38e874