[TASK] Improve caching and loading of VIE and node type schemata
This change implements a client-side resource cache that uses the
session storage and allows for early preloading of resources before
the page is loaded which speeds up the UI loading process.

Additionally the Neos node type schema is not embedded into the markup
but also loaded via the same mechanism as the VIE schema.

Change-Id: I63b663cc8a1ae62bfd1da4e6837fc676ee56a98b
Resolves: #44976