[BUGFIX] Make user menu links clickable in content module
When clicking a link inside the user menu the menu was closed
and the underlying element got clicked instead.

This was due to the document having a click event being
triggered that closed all dropdown menus, because click
events inside the dropdown menu propagated up.

The issue has been fixed in the Twitter Bootstrap script,
so this is just an update of that script.

Change-Id: I1b9fe015b6c4865eef35ebf20e0fe3554f15a22d