[TASK] Update TypoScript, FlowQuery, Eel documentation
This change updates the TypoScript part of the Integrators Guide to make
it more consistent and up-to-date with recent changes. It also fills a
number of holes in it: TS namespaces, processors, Eel, FlowQuery, Fizzle
are now documented.

A FlowQuery operations reference is added, generated using TYPO3.DocTools
with a new parser that has been added.

Change-Id: Ice3b94c505a590baa69c5e445f6fb2e6280464ce
Related: #45004