[!!!][TASK] Remove deprecated Aloha (Not)Editable view helpers
The Aloha Editable and NotEditableViewHelper have been deprecated since
sprint 10 and are removed with this change.

The NotEditable view helper can be removed without replacement in
templates. The Editable view helper has been replaced with the
contentElement.editable view helper. It has the same signature, so that
adjustments besides renaming it in template are not needed.

Change-Id: I606b5b3fd604a623a59afb42a0411ccf7a4efbdd