[TASK] Move backend elements into one container element
Change the Ember.Application initialization so all events
and application handling is really bound to the container
div element. Also adds the major UI components in the
correct way to the application template instead of
using $.appendTo().

Replacing jquery.lionbars with antiscroll, fixes some
style issues and removes the launcher which does
not fit in the masterplan.

Change-Id: I5fad06a80bb87df6fb4cfec086c96e6d864cc022