[FEATURE] show Aloha in Top Bar and Inspector
Splits up Aloha functionality, such that the formatting tab is shown
in the top bar, and the remaining Aloha functionality is shown in the
inspector; rendered with the Neos editing controls.

This is not yet styled, but it should work.

This change furthermore renames "contentCollectionsAndViews" to
"groupedPropertyViews" in Controller.Inspector. This was renamed accidentally
during the "section -> contentCollection" renaming.

Resolves: #48980
Change-Id: I05afdb89bc02707a2f4f74e6948f24e5e502511d