[FEATURE] Integrate Mediabrowser into Neos, get rid of popovers
Make sure to add the corresponding change to TYPO3.Media as well.

Instead of popovers, we now use sliders for HTML editor, the media
browser and the cropping.

The secondary inspector is not yet fully styled, but already works
functionality-wise.

Furthermore, this change removes all popover styles from CSS.

Change-Id: Idc89d47cb2e55098efd151d848fd0b8972f30f5c
Related: #45003
Reviewed-on: https://review.typo3.org/21848
Reviewed-by: Sebastian Kurfuerst
Tested-by: Sebastian Kurfuerst