[!!!][FEATURE] Restrict access to specific modules
Additionally adds a version parameter to the menu
controller request, to ensure that changes in roles
will update the accessible modules accordingly.

This is breaking since it's no longer possible to use
the standard module controller if no controller is
specified. This is done to be able to enforce restrictions
on overview modules with resource policies.

Change-Id: Iad2207afc5c64a9e66d8785a6fa2e70217dc0ccb
Resolves: #49949
Reviewed-on: https://review.typo3.org/22131
Reviewed-by: Christopher Hlubek
Tested-by: Christopher Hlubek
Tested-by: Mr. Hudson
Reviewed-by: Aske Ertmann
Tested-by: Aske Ertmann