[BUGFIX] Fix menu and backend URI generation for multiple sites
This change allows to have multiple sites and enables linking to them
if an active domain is present on a site. The cached resource is
updated whenever a Site or Domain is updated through the configuration
cache version.

A site without an active domain will be shown as disabled in the menu,
because we cannot link to it (for now).

Change-Id: Ia487671edd64da0cfbadf591dd3c4961d18e9132
Related: #48068
Reviewed-on: https://review.typo3.org/22120
Reviewed-by: Aske Ertmann
Tested-by: Aske Ertmann
Reviewed-by: Christopher Hlubek
Tested-by: Christopher Hlubek