[TASK] Add T3.Common.Notification as a shared resource
This adds the T3.Common.Notification as a shared resource.
Also replaces the ocurrances of the old usage in the new
JavaScript structure.

But it do not change the PageTree because of the refactoring
in the Navigate component (I393fd01266f95ece643999b4e8e59e3710dc76f5)
commit.

Change-Id: I8f14ae178c056087878a39be565db1dd5b0d57a4
Related: #48328
Reviewed-on: https://review.typo3.org/23491
Reviewed-by: Aske Ertmann
Tested-by: Aske Ertmann