[TASK] Linking to Nodes with given ContextNodePath
Giving a ContextNodePath that is an absolute path appended
with the context, like "/sites/flowtypo3org/homepage@live",
correctly creates a link for this node, even without being
inside a Neos environment and, therefore, without having
access to the TypoScript environment which contains
Node objects.

This makes interoperability between Flow applications and Neos
easier as it allows you to render links from Flow applications
without the needing to create a Node object.

Change-Id: Iad62542b619a80534f08967044c84a637f4ec7f0
Reviewed-on: https://review.typo3.org/22262
Reviewed-by: Sebastian Kurfuerst
Tested-by: Sebastian Kurfuerst