[BUGFIX] Graceful degradation of broken content elements
If content elements cannot be rendered for some reason, an error message
is displayed instead of the actual content. The broken elements can
still be (un)hidden, copy-pasted, … and even published.

Some errors are not yet covered yet, like not existing Node-Types.
In this case the error message is displayed without the option to hide,
remove, cut or copy the broken element.
There is already a request for this Feature #52511.

The css-styles of the Neos backend do not apply to published, broken
content elements, but can be customized.

The TypoScript object type TYPO3.Neos:Template declares the
@exceptionHandler property to handle rendering exceptions with a special
handler that wraps nodes with the content element wrapping services.

Resolves: #45531
Change-Id: Iddd7b254e352b7960919bfc0c25321e2a38660e4
Reviewed-on: https://review.typo3.org/24239
Reviewed-by: Christopher Hlubek
Tested-by: Christopher Hlubek
Reviewed-by: Sebastian Kurfuerst
Tested-by: Sebastian Kurfuerst