[TASK] Clarify pasting of nodes
Previously the paste button always said "Paste after" even though
it actually pastes into a content collection. This change
introduces a new title property, so we actually have the correct
one depending on the context and also removes the "after" from all
methods and parameters as this is handled in the ContentCommands.

Change-Id: Ifff23d9bf4d6204c4671c360547754a1ee3b9c57
Relates: #40677
Reviewed-on: https://review.typo3.org/24293
Reviewed-by: Tim Kandel
Tested-by: Tim Kandel
Tested-by: Mr. Hudson
Reviewed-by: Oliver Eglseder
Reviewed-by: Rens Admiraal
Tested-by: Rens Admiraal