[!!!][FEATURE] improve reloading of content without neos-reloadable-content
This change removes the neos-reloadable-content CSS class, which was
used to determine the sections which shall be reloaded. Now,
we reload everything between #t3-page-metainformation (which is
rendered directly after the body tag by the Neos PageTemplate.html)
and #neos-application, which is rendered at the very end of the body
by the same template.

This greatly improves the reloading experience, taking also changes
to menu states, breadcrumb, and complete layout modifications into account.

Applying new CSS classes on <html> and <body> is not yet supported,
though (but that did not work before this change, either).

This change is breaking because <neos:contentElement> does not have
a reloadable-attribute anymore. Just remove these classes from your
usages of <neos:contentElement>.

Make sure to check the corresponding changes as well.

Change-Id: Ic2b2c1f6cdf21c8bd251426b14d877bdccfbbe7e
Resolves: #49945
Reviewed-on: https://review.typo3.org/24300
Reviewed-by: Aske Ertmann
Tested-by: Aske Ertmann