[!!!][TASK] Remove automagic creation of ContentCollections on pages
ContentCollections are no longer auto created on rendering of a
page. Instead they need to be defined as childNodes of the Page
Node. Neos predefines a "main" section by default which can be
overriden or added to in the site package. Additionally you can
of course create your own Document Type with a specific child
node structure.

This is breaking if you use more than the "main" section in your
site. In this case you have to define those additional sections
in the NodeTypes.yaml of your site package.

Related: #49951
Change-Id: I1017c59d89a67447667d7a3ed0dc3bd92cd6dd0d
Reviewed-on: https://review.typo3.org/24338
Reviewed-by: Christian Mueller
Tested-by: Christian Mueller