[TASK] Disallow access to foreign workspaces
The workspace module currently allows
to show, publish and discard workspaces
of other users.
Since we currently only have per user workspaces,
this change disables the selection of other
workspaces in the workspaces module controller
so that only access to the workspace of
the currently logged in user is possible.

Resolves: #52506
Resolves: #52504
Change-Id: I01be491ba5bd415f861fc459e7c5e0bb74c897f6
Reviewed-on: https://review.typo3.org/24342
Reviewed-by: Andreas Förthner
Tested-by: Andreas Förthner