[!!!][TASK] Make TYPO3.Neos.nodeTypes.groups extensible and sortable
This changes allows package authors to define their own nodeTypeGroups
without having to overwrite the default settings. It also gets rid of the need
for the integrator to redeclare every nodeTypeGroup in the Distributions
Settings.yaml if he wants to add new ones.

The configuration of TYPO3.Neos.nodeTypes.groups changes to the following format::

  TYPO3:
    Neos:
      nodeTypes:
        groups:
          general:
            position: 'start'
            label: 'General'
          structure:
            position: 100
            label: 'Structure'

This change is breaking because the "group" settings in the NodeTypes.yaml
now has to use the key instead of the label and the Settings.yaml
structure has changed.

Change-Id: I61212ebc08b4824f6e8be7a1b6a60207fc98e40b
Resolves: #52524
Reviewed-on: https://review.typo3.org/24356
Reviewed-by: Karsten Dambekalns
Tested-by: Karsten Dambekalns