[BUGFIX] The 404 error page is rendered outside viewport
The 404 page is styled incorrectly and because of this not
displayed inside the screen, but only partly visible in the top
right corner of the screen.

This change cleans up the dependencies of the error screen on
bootstrap and minifies the CSS used. Besides that it adds
responsive styling for the error page to display it in the
center of the page again.

Change-Id: I3736cb102da5cbab46e27da26d66439cd5a371d2
Resolves: #52711
Reviewed-on: https://review.typo3.org/24920
Reviewed-by: Aske Ertmann
Tested-by: Aske Ertmann
Tested-by: Mr. Hudson
Reviewed-by: Rens Admiraal
Tested-by: Rens Admiraal