[BUGFIX] Click events are caught in the entity selection
Currently we catch all click events on the document to update
the selected node and prevent the click event to pass through
to other listeners.

One issue that this causes is that it's not possible to use the
middle mouse click to open links in a new tab in the menu e.g.

Change-Id: I0a09dc765cf3bf4761828a6e751f95360d849b1e
Reviewed-on: https://review.typo3.org/24969
Reviewed-by: Aske Ertmann
Tested-by: Aske Ertmann
Reviewed-by: Dominique Feyer
Tested-by: Dominique Feyer
Reviewed-by: Rens Admiraal
Tested-by: Rens Admiraal
Tested-by: Mr. Hudson
Reviewed-by: Sebastian Kurfuerst
Tested-by: Sebastian Kurfuerst