[BUGFIX] Fix required argument exception when copying a node
This fixes an Exception that occures when copying a node
as the argument nodeName is not set.
Besides that it fixes the redirect after the node is pasted.
The redirect was not taking the nextUri into account. As
nextUri seemed to be incorrectly used in the AbstractNodeTree
a nodeUri property is added to the request.

Change-Id: Ibf994d26c1224edf04994291d74f8943b7582dfc
Reviewed-on: https://review.typo3.org/25081
Reviewed-by: Aske Ertmann
Tested-by: Aske Ertmann
Tested-by: Mr. Hudson
Reviewed-by: Rens Admiraal
Tested-by: Rens Admiraal