[BUGFIX] Element handles and overlay are mispositioned
In some cases, if the browser window size changes, or the
website style has some transitions which are run, the
positioning of the handles is not updated.

This change introduces a helper that, using a setInterval(),
allows for actions to be done continuosly. Currently the helper
forces the width of the overlay of not inline editable elements
and the position of the content element handles.

Change-Id: I816399b6ddb37b2c0515ab32296d0edf3fb89b36
Reviewed-on: https://review.typo3.org/25126
Reviewed-by: Sebastian Kurfuerst
Tested-by: Sebastian Kurfuerst
Tested-by: Mr. Hudson
Reviewed-by: Tim Kandel
Tested-by: Tim Kandel