[!!!][FEATURE] Auto-generate TypoScript from node type configuration
This change refactors the TypoScriptService for easier testing in the
RenderingTest and adds the feature of auto-generating TypoScript from the
node type configuration.

The generated code for node types is the following::

    prototype(TYPO3.Neos.NodeTypes:Text) < prototype(TYPO3.Neos:Template) {
        templatePath = 'resource://TYPO3.Neos.NodeTypes/Private/Templates/NodeTypes/Text.html'
        text = ${node.properties.text}
    }

This is non-breaking as long as the prototype for a node type does not
use Array or another TypoScript object, where the additional properties
of the auto-generated code could produce unwanted output during rendering.

For these cases it is sufficient to reset the prototype definition in
front of the custom declaration::

    prototype(TYPO3.Neos:Page) >
    prototype(TYPO3.Neos:Page) < prototype(TYPO3.TypoScript:Array) {
        ...
    }

Due to the changed way of merging the TypoScript relative paths in
"include" statements might need to be changed. Relative paths are
always resolved from the directory of the current file.

Change-Id: If89c54a3e50f6c17f60b968abeb146daa1659bc4
Resolves: #53321
Reviewed-on: https://review.typo3.org/25161
Reviewed-by: Christopher Hlubek
Tested-by: Christopher Hlubek