[FEATURE] Support tag name and attributes for ContentElementViewHelper
This change extends the neos:contentElement view helper that is used
to wrap nodes with RDFa and editing metadata.

It adds a configurable tag name and registers universal tag attributes
to override all attributes of the wrapping tag from the view helper.

This is the first step to reduce the generated code and nesting by
allowing more fine-grained control over the markup.

Change-Id: Iadb881d9545d442b6816ec105c75ac62bd99cac2
Related: #53046
Reviewed-on: https://review.typo3.org/25174
Reviewed-by: Christopher Hlubek
Tested-by: Christopher Hlubek
Reviewed-by: Aske Ertmann
Tested-by: Mr. Hudson