[!!!][FEATURE] Support custom tag name and attributes for ContentCollection
This change adds new TypoScript properties to the ContentCollection
object to set the tagName and attributes of the generated tag that wraps
the collection.

Example::

    prototype(Acme.Demo:List) < prototype(TYPO3.Neos:Template) {
        items = TYPO3.Neos:ContentCollection {
            nodePath = 'items'
            tagName = 'ul'
            attributes.class = 'my-list'
        }
    }

The TypoScript code will render the following HTML::

    <ul class="neos-contentcollection my-list">
        ...
    </ul>

This is a breaking change as the "className" property for a custom class
name on the generated tag is removed. It has to be changed to the
"attributes.class" declaration.

Change-Id: Ia006e962eaaafb35d4e362900b093e969c9ebb9d
Related: #53046
Reviewed-on: https://review.typo3.org/25179
Tested-by: Christopher Hlubek
Tested-by: Mr. Hudson
Reviewed-by: Aske Ertmann
Reviewed-by: Rens Admiraal
Tested-by: Rens Admiraal
Reviewed-by: Christopher Hlubek