[!!!][FEATURE] Use node reference for Shortcut target node
This changes the property type of the target node for Shortcut pages
from a simple node path to a real node reference.

If the target node property contains a path, it is silently ignored.

Also fixes the bug which only showed the first child node if the target
mode was switched forth and back and saved in between.

Change-Id: I67ea6fa3d282a40c7c661b41a97799fef0b8a43b
Reviewed-on: https://review.typo3.org/25202
Reviewed-by: Aske Ertmann
Tested-by: Aske Ertmann
Reviewed-by: Robert Lemke
Tested-by: Robert Lemke