[FEATURE] When exporting a site, create files for resources
Instead of only exporting XML to standard output it can now be written
to a file directly. If this is done, any resources will no longer be
embedded into the XML (as base64 encoded data) but will be placed into
a folder alongside the XML file.

This massively reduces the size of the generated XML.

Change-Id: Icb716fa5f8b450e7303a2bb4555973927b9fcfc3
Reviewed-on: https://review.typo3.org/25241
Reviewed-by: Robert Lemke
Tested-by: Robert Lemke