[BUGFIX] Exporting a site sets state to empty string
The state of a site is an integer, but since the state is exported like
a node property since Ifa3b6cb89a2b8ed10eaddaf6f450b98b8c96fe98 that
failed because all properties are expected to be strings.

This change exports any scalar property by typecasting it to string.

Change-Id: I906b3b774aba2c560f7c0642af0c9082850e224a
Reviewed-on: https://review.typo3.org/25385
Tested-by: Karsten Dambekalns
Reviewed-by: Bastian Waidelich
Tested-by: Mr. Hudson
Reviewed-by: Dominique Feyer
Tested-by: Dominique Feyer
Reviewed-by: Karsten Dambekalns