[FEATURE] Edit/Preview Panel
Introduces the positibility to add custom edit and preview modes
that can then be selected by the editor. A mode can alter the
rendering to allow a different presentation of the content.

This also separates the preview mode from the full screen mode,
making all editing and preview modes available in full screen.

Various improvements to the raw content mode is included as well,
inclusing renaming from wireframe, templates and supporting shortcuts.

Change-Id: I3c6e42d3f3e38770bb8d41d5e306929972ad00d7
Reviewed-on: https://review.typo3.org/24949
Reviewed-by: Aske Ertmann
Tested-by: Aske Ertmann
Reviewed-by: Tim Kandel
Tested-by: Tim Kandel
Reviewed-by: Sebastian Helzle
Tested-by: Sebastian Helzle
Tested-by: Mr. Hudson
Reviewed-by: Christopher Hlubek
Tested-by: Christopher Hlubek