[BUGFIX] Fix always-activated Publish button
Currently, if using the Neos Content Module, there is always
exactly one change even if nothing was modified, always activating
the "publish" button.

This has been traced down to a regression introduced by
I26517fd0506cc1074b90e84e5a465001421ae36c, which set the
workspace of the current node always to the current user
workspace (which does not make sense). The workspace of
the *node* must be the "real" workspace this node is in,
and the workspace of the *node's context* is the one the
user is currently inside.

Change-Id: Ib460932437415a0fd2f2e7fc2a7d8fda1db4d579
Reviewed-on: https://review.typo3.org/25449
Reviewed-by: Sebastian Kurfuerst
Tested-by: Sebastian Kurfuerst
Tested-by: Mr. Hudson
Reviewed-by: Dominique Feyer
Tested-by: Dominique Feyer
Reviewed-by: Karsten Dambekalns
Tested-by: Karsten Dambekalns