[TASK] Add filter by node for the filter operation
The FlowQuery operation is now able to filter by a given node, this
filter type always return a context with a maximum of one node.

This FlowQuery return one if the currentNode is a parent node of node::

${q(node).parents(currentNode).count()}

This operation can be used in the MenuImplementation to test if the
current node is a child node of a given node, to set the active and
current class.

Change-Id: I71c72d8fdc6d2b36b11ed3fdac49affaffd9ca8c
Reviewed-on: https://review.typo3.org/25561
Reviewed-by: Dominique Feyer
Tested-by: Dominique Feyer
Reviewed-by: Aske Ertmann
Reviewed-by: Christopher Hlubek
Tested-by: Christopher Hlubek